providerを導入したが画面が真っ白?
