+++ pubspec.yaml
@@ -36,6 +36,7 @@ dependencies:
   cupertino_icons: ^1.0.2
   firebase_core: ^1.14.0
   cloud_firestore: ^3.1.11
+  provider: ^6.0.2
 
 dev_dependencies:
   flutter_test:
